+++ p2panda-rs/src/document/document_view_fields.rs
@@ -117,7 +117,7 @@ mod tests {
     use crate::document::{DocumentViewFields, DocumentViewValue};
     use crate::operation::traits::AsOperation;
     use crate::operation::{OperationId, OperationValue};
-    use crate::test_utils::db::PublishedOperation;
+    use crate::test_utils::memory_store::PublishedOperation;
     use crate::test_utils::fixtures::{published_operation, random_operation_id};
     use crate::WithId;
