+++ p2panda-rs/src/test_utils/fixtures/operation_fixtures.rs
@@ -16,7 +16,7 @@ use crate::operation::{
 };
 use crate::schema::{Schema, SchemaId};
 use crate::test_utils::constants::{test_fields, HASH, SCHEMA_ID};
-use crate::test_utils::db::PublishedOperation;
+use crate::test_utils::memory_store::PublishedOperation;
 use crate::test_utils::fixtures::{
     document_view_id, key_pair, random_document_id, random_hash, schema, schema_fields, schema_id,
 };
