+++ p2panda-rs/src/test_utils/memory_store/mod.rs
@@ -1,12 +1,12 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-//! Implementations of all storage provider traits.
+//! `MemoryStore` implementation of all storage provider traits.
 //!
 //! Used in the mock node and for testing.
 pub mod domain;
 mod provider;
 mod stores;
-pub mod test_db;
+pub mod helpers;
 mod types;
 pub mod validation;
